made modal into component
