+++ src/components/ProfileSwitchModal/styles.tsx
@@ -1,0 +1,111 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  modalImage: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    width: 40,
+    height: 40,
+    position: 'absolute',
+    left: '8%',
+  },
+  switchProfileModalView: {
+    position: 'absolute',
+    width: '87%',
+    height: '46%',
+    top: '20%',
+    left: '6%',
+    backgroundColor: '#FFFFFF',
+    borderRadius: 10,
+    borderWidth: 2,
+    borderColor: '#F2F2F2',
+  },
+  switchModalX: {
+    display: 'flex',
+    position: 'absolute',
+    top: '5%',
+    right: '5%',
+  },
+  flatListItem: {
+    display: 'flex',
+    alignContent: 'center',
+    justifyContent: 'center',
+    width: '100%',
+    marginVertical: 15,
+  },
+  headFlatListItem: {
+    display: 'flex',
+    alignContent: 'center',
+    justifyContent: 'center',
+    width: '100%',
+    top: '16%',
+  },
+  listInfo: {
+    display: 'flex',
+    left: '25%',
+  },
+  modalDivider: {
+    position: 'absolute',
+    height: '0%',
+    width: '86%',
+    borderStyle: 'solid',
+    borderColor: '#F2F2F2',
+    top: '28%',
+    left: '6%',
+    borderWidth: 1,
+  },
+  itemSeparator: {
+    height: '0%',
+    width: '68%',
+    borderStyle: 'solid',
+    borderColor: '#F2F2F2',
+    borderWidth: 1,
+    left: '25%',
+  },
+  signedIn: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#F2F2F2',
+    width: '23%',
+    height: 24,
+    position: 'absolute',
+    right: '7%',
+    borderRadius: 26,
+  },
+  switch: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    height: 24,
+    width: '18%',
+    position: 'absolute',
+    right: '7%',
+    backgroundColor: '#F2F2F2',
+    borderRadius: 26,
+  },
+  flatList: {
+    position: 'absolute',
+    top: '29%',
+    width: '100%',
+    height: '70%',
+  },
+  addPressable: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#D9D9D9',
+    width: 30,
+    height: 30,
+    position: 'absolute',
+    left: '8.8%',
+    borderRadius: 50,
+  },
+  addMember: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignContent: 'center',
+    marginTop: 18,
+  },
+});
