+++ src/screens/Profile/ProfileScreen.tsx
@@ -17,111 +17,22 @@ import ViewContainer from "../../components/ViewContainer";
 import styles from "./styles";
 import SvgIcon from "../../../assets/SvgIcon";
 import globalStyles from "../../globalStyles";
-
-function Item({ name, role, status }) {
-  if (status == 'Signed In') {
-    return (
-      <View style={styles.headFlatListItem}>
-        <View style={styles.photoFiller} />
-        <View style={styles.listInfo}>
-          <Text style={globalStyles.body1Bold}>{name}</Text>
-          <Text style={globalStyles.body2}>{role}</Text>
-        </View>
-        <Pressable style={styles.signedIn}>
-          <Text style={globalStyles.body2}>{status}</Text>
-        </Pressable>
-      </View>
-    );
-  }
-  return (
-    <View style={styles.flatListItem}>
-      <View style={styles.photoFiller} />
-      <View style={styles.listInfo}>
-        <Text style={globalStyles.body1Bold}>{name}</Text>
-        <Text style={globalStyles.body2}>{role}</Text>
-      </View>
-      <Pressable style={styles.switch}>
-        <Text style={globalStyles.body2}>{status}</Text>
-      </Pressable>
-    </View>
-  );
-}
+import ProfileSwitchModal from "../../components/ProfileSwitchModal/ProfileSwitchModal";
 
 function ProfileScreen() {
   const [profileSwitchModalVisible, setProfileSwitchModalVisible] =    useState(false);
 
-  const DATA = [
-    {
-      id: "1",
-      name: "Kelly Hu",
-      role: "Dependent",
-    },
-    {
-      id: "2",
-      name: "Albert Liu",
-      role: "Dependent",
-    },
-    {
-      id: "3",
-      name: "Sydney Bui",
-      role: "Dependent",
-    },
-    // {
-    //   id: '4',
-    //   name: 'Gabe Espinosa',
-    //   role: 'Head',
-    // },
-    // {
-    //   id: '5',
-    //   name: 'Danashi Messele',
-    //   role: 'Dependent',
-    // },
-  ];
-
-  const renderItem = ({ item }) => (
-    <Item name={item.name} role={item.role} status="Switch" />
-  );
-
-  const itemSeparator = () => <View style={styles.itemSeparator} />;
-
-  const addMember = () => (
-    <View style={styles.addMember}>
-      <Pressable style={styles.addPressable}>
-        <SvgIcon type="addPlus" />
-      </Pressable>
-      <Text style={[globalStyles.body1Bold, styles.listInfo]}>Add Member</Text>
-    </View>
-  );
-
   return (
-    <View
-      style={
-        profileSwitchModalVisible
-          ? styles.modalVisibleContainer
-          : globalStyles.container
-      }
-    >
+    <ViewContainer>
+      <View
+        style={profileSwitchModalVisible ? styles.modalVisibleContainer : null}
+      />
 
-      <Modal visible={profileSwitchModalVisible} transparent>
-        <View style={styles.switchProfileModalView}>
-          <Pressable
-            style={styles.switchModalX}
-            onPress={() => setProfileSwitchModalVisible(false)}
-          >
-            <SvgIcon type="greyX" />
-          </Pressable>
-          <Item name="Jacob Kim" role="Head" status="Signed In" />
-          <View style={styles.modalDivider} />
-          <FlatList
-            style={styles.flatList}
-            data={DATA}
-            renderItem={renderItem}
-            keyExtractor={(item) => item.id}
-            ItemSeparatorComponent={itemSeparator}
-            ListFooterComponent={addMember}
-          />
-        </View>
-      </Modal>
+      <ProfileSwitchModal
+        visible={profileSwitchModalVisible}
+        setVisible={setProfileSwitchModalVisible}
+        userID="userIDfiller"
+      />
 
       <Pressable
         style={styles.profileImagePressable}
@@ -157,11 +68,11 @@ function ProfileScreen() {
         <Text style={globalStyles.body1}>booblywobbly@gmail.com</Text>
         <Text style={globalStyles.body1}>********</Text>
       </View>
-      
+
       <Pressable style={styles.logoutPressable}>
         <Text style={globalStyles.overline2}>LOGOUT</Text>
       </Pressable>
-    </View>
+    </ViewContainer>
   );
 }
