+++ src/screens/Profile/styles.tsx
@@ -1,198 +1,92 @@
-import { StyleSheet } from 'react-native';
+import { StyleSheet } from "react-native";
 
 export default StyleSheet.create({
   profileName: {
-    position: 'absolute',
-    top: '24%',
-    textAlign: 'center',
+    position: "absolute",
+    top: "24%",
+    textAlign: "center",
   },
   profileImage: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignContent: 'center',
-  },
-  editPressable: {
-    display: 'flex',
-    position: 'absolute',
-    alignItems: 'center',
-    justifyContent: 'center',
-    width: '40%',
-    height: '5%',
-    top: '30%',
-    left: '8%',
-    backgroundColor: '#F2F2F2',
-  },
-  resetPressable: {
-    display: 'flex',
-    position: 'absolute',
-    top: '30%',
-    right: '8%',
-    alignItems: 'center',
-    justifyContent: 'center',
-    width: '40%',
-    height: '5%',
-    backgroundColor: '#F2F2F2',
-  },
-  profileFields: {
-    display: 'flex',
-    flexDirection: 'column',
-    position: 'absolute',
-    left: '8%',
-    top: '39.6%',
-    height: '23%',
-    justifyContent: 'space-between',
-  },
-  profileValues: {
-    display: 'flex',
-    flexDirection: 'column',
-    position: 'absolute',
-    left: '37%',
-    top: '39.6%',
-    height: '23%',
-    justifyContent: 'space-between',
-  },
-  logoutPressable: {
-    display: 'flex',
-    position: 'absolute',
-    alignItems: 'center',
-    justifyContent: 'center',
-    top: '77%',
-    width: '40%',
-    height: '5%',
-    backgroundColor: '#C4C4C4',
-  },
-  profileImagePressable: {
-    display: 'flex',
-    position: 'absolute',
-    alignItems: 'center',
-    justifyContent: 'center',
-    top: '6%',
-    width: '37%',
-    height: '16%',
-    backgroundColor: '#FFFFFF',
-    borderRadius: 10,
-    borderColor: '#F2F2F2',
-    borderWidth: 2,
-    borderStyle: 'solid',
-  },
-  switchProfileModalView: {
-    position: 'absolute',
-    width: '87%',
-    height: '46%',
-    top: '20%',
-    left: '6%',
-    backgroundColor: '#FFFFFF',
-    borderRadius: 10,
-    borderWidth: 2,
-    borderColor: "#F2F2F2",
-  },
-  switchModalX: {
-    display: 'flex',
-    position: 'absolute',
-    top: '5%',
-    right: '5%',
-  },
-  flatListItem: {
-    display: 'flex',
-    alignContent: 'center',
-    justifyContent: "center",
-    width: '100%',
-    marginVertical: 15,
-  },
-  headFlatListItem: {
-    display: 'flex',
-    alignContent: 'center',
-    justifyContent: "center",
-    width: '100%',
-    top: "15%",
-  },
-  listInfo: {
-    display: 'flex',
-    left: '25%',
-  },
-  modalDivider: {
-    position: 'absolute',
-    height: '0%',
-    width: '86%',
-    borderStyle: 'solid',
-    borderColor: '#F2F2F2',
-    top: "28%",
-    left: '6%',
-    borderWidth: 1,
-  },
-  itemSeparator: {
-    height: '0%',
-    width: '68%',
-    borderStyle: 'solid',
-    borderColor: '#F2F2F2',
-    borderWidth: 1,
-    left: "25%",
-  },
-  signedIn: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F2F2F2',
-    width: "23%",
-    height: 24,
-    position: 'absolute',
-    right: '7%',
-    borderRadius: 26,
-  },
-  switch: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    height: 24,
-    width: '18%',
-    position: 'absolute',
-    right: '7%',
-    backgroundColor: '#F2F2F2',
-    borderRadius: 26,
-  },
-  flatList: {
-    position: 'absolute',
-    top: '29%',
-    width: '100%',
-    height: '70%',
-  },
-  addPressable: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#D9D9D9',
-    width: 30,
-    height: 30,
-    position: 'absolute',
-    left: '8.8%',
-    borderRadius: 50,
-  },
-  addMember: {
     display: "flex",
     justifyContent: "center",
     alignContent: "center",
-    marginTop: 18,
   },
-  photoFiller: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#D9D9D9',
-    width: 35,
-    height: 35,
-    position: 'absolute',
-    left: '8%',
-    borderRadius: 50,
-  },
-  downArrow: {
+  editPressable: {
+    display: "flex",
     position: "absolute",
-    bottom: "6%",
+    alignItems: "center",
+    justifyContent: "center",
+    width: "40%",
+    height: "5%",
+    top: "30%",
+    left: "8%",
+    backgroundColor: "#F2F2F2",
+  },
+  resetPressable: {
+    display: "flex",
+    position: "absolute",
+    top: "30%",
+    right: "8%",
+    alignItems: "center",
+    justifyContent: "center",
+    width: "40%",
+    height: "5%",
+    backgroundColor: "#F2F2F2",
+  },
+  profileFields: {
+    display: "flex",
+    flexDirection: "column",
+    position: "absolute",
+    left: "8%",
+    top: "39.6%",
+    height: "23%",
+    justifyContent: "space-between",
+  },
+  profileValues: {
+    display: "flex",
+    flexDirection: "column",
+    position: "absolute",
+    left: "37%",
+    top: "39.6%",
+    height: "23%",
+    justifyContent: "space-between",
+  },
+  logoutPressable: {
+    display: "flex",
+    position: "absolute",
+    alignItems: "center",
+    justifyContent: "center",
+    top: "77%",
+    width: "40%",
+    height: "5%",
+    backgroundColor: "#C4C4C4",
+  },
+  profileImagePressable: {
+    display: "flex",
+    position: "absolute",
+    alignItems: "center",
+    justifyContent: "center",
+    top: "6%",
+    width: "37%",
+    height: "16%",
+    backgroundColor: "#FFFFFF",
+    borderRadius: 10,
+    borderColor: "#F2F2F2",
+    borderWidth: 2,
+    borderStyle: "solid",
   },
   modalVisibleContainer: {
     flex: 1,
-    backgroundColor: '#FAFAFA',
+    backgroundColor: '#000000',
+    width: '100%',
+    height: '100%',
     opacity: 0.4,
     alignItems: 'center',
     justifyContent: 'center',
+    zIndex: 10,
+  },
+  downArrow: {
+    position: 'absolute',
+    bottom: '6%',
   },
 });
